+++ mml/src/message/body/compiler/parsers/parts.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use crate::message::body::{
     compiler::tokens::{Part, Props},
-    GREATER_THAN, MULTI_PART_BEGIN, MULTI_PART_END,
+    GREATER_THAN, MULTIPART_BEGIN, MULTIPART_END,
 };
 
 use super::{description, disposition, filename, multipart_type, name, part_type, prelude::*};
@@ -13,43 +13,54 @@ use super::{encrypt, sign};
 ///
 /// It parses all parts the MML body is composed of.
 pub(crate) fn parts<'a>() -> impl Parser<'a, &'a str, Vec<Part<'a>>, ParserError<'a>> + Clone {
-    choice((
-        multi_part(),
-        single_part(),
-        plain_text_part(1).map(Part::PlainText),
-    ))
-    .repeated()
-    .collect()
-    .then_ignore(end())
-}
-
-/// The plain text part parser.
-///
-/// It parses everything that is inside and outside (multi)parts.
-pub(crate) fn plain_text_part<'a>(
-    min: usize,
-) -> impl Parser<'a, &'a str, &'a str, ParserError<'a>> + Clone {
-    any()
-        .and_is(
-            choice((
-                single_part_begin(),
-                single_part_end(),
-                multi_part_begin(),
-                multi_part_end(),
-            ))
-            .not(),
-        )
+    choice((multipart(), part(), plain_text_part(1).map(Part::PlainText)))
         .repeated()
-        .at_least(min)
-        .slice()
+        .collect()
+        .then_ignore(end())
 }
 
-/// The single part parser.
+/// The multipart parser.
 ///
-/// It parses a full part, including properties and content till the
-/// next opening part or the next closing part/multipart.
-pub(crate) fn single_part<'a>() -> impl Parser<'a, &'a str, Part<'a>, ParserError<'a>> + Clone {
-    single_part_begin()
+/// It parses everything between tags `<#multipart>` and
+/// `<#/multipart>`. A multipart can contain multiple parts as well as
+/// multiple other multiparts (recursively). This parser is useful
+/// when you need to group parts together instead of having them at
+/// the root level.
+pub(crate) fn multipart<'a>() -> impl Parser<'a, &'a str, Part<'a>, ParserError<'a>> + Clone {
+    recursive(|multipart| {
+        just(MULTIPART_BEGIN)
+            .ignore_then(
+                choice((
+                    multipart_type(),
+                    description(),
+                    #[cfg(feature = "pgp")]
+                    encrypt(),
+                    #[cfg(feature = "pgp")]
+                    sign(),
+                ))
+                .repeated()
+                .collect::<Props>(),
+            )
+            .then_ignore(just(GREATER_THAN))
+            .then_ignore(new_line().or_not())
+            .then(
+                choice((multipart, part(), plain_text_part(1).map(Part::PlainText)))
+                    .repeated()
+                    .collect(),
+            )
+            .then_ignore(just(MULTIPART_END))
+            .then_ignore(new_line().or_not())
+            .map(|(props, parts)| Part::Multi(props, parts))
+    })
+}
+
+/// The part parser.
+///
+/// It parses a full part, including properties and contents from
+/// `<#part>` till the next opening part or the next closing
+/// `<#/part>` or `<#/multipart>`.
+pub(crate) fn part<'a>() -> impl Parser<'a, &'a str, Part<'a>, ParserError<'a>> + Clone {
+    part_begin()
         .ignore_then(
             choice((
                 part_type(),
@@ -68,65 +79,21 @@ pub(crate) fn single_part<'a>() -> impl Parser<'a, &'a str, Part<'a>, ParserErro
             .then_ignore(new_line().or_not()),
         )
         .then(plain_text_part(0))
-        .then_ignore(single_part_end().then(new_line().or_not()).or_not())
+        .then_ignore(part_end().then(new_line().or_not()).or_not())
         .map(|(props, content)| Part::Single(props, content))
 }
 
-/// The multipart parser.
+/// The plain text part parser.
 ///
-/// It parses everything between tags `<#multipart>` and
-/// `<#/multipart>`. A multipart can contain multiple parts as well as
-/// multiple other multiparts (recursively). This parser is useful
-/// when you need to group parts together instead of having them at
-/// the root level.
-///
-/// # Examples
-///
-/// From https://www.gnu.org/software/emacs/manual/html_node/emacs-mime/Advanced-MML-Example.html:
-///
-/// ```ignore
-/// // <#multipart type=mixed>
-/// //   <#part type=image/jpeg filename=~/rms.jpg disposition=inline>
-/// //   <#multipart type=alternative>
-/// //     This is a plain text part.
-/// //     <#part type=text/enriched name=enriched.txt>
-/// //     <center>This is a centered enriched part</center>
-/// //   <#/multipart>
-/// //   This is a new plain text part.
-/// //   <#part disposition=attachment>
-/// //   This plain text part is an attachment.
-/// // <#/multipart>
-/// ```
-pub(crate) fn multi_part<'a>() -> impl Parser<'a, &'a str, Part<'a>, ParserError<'a>> + Clone {
-    recursive(|multipart| {
-        just(MULTI_PART_BEGIN)
-            .ignore_then(
-                choice((
-                    multipart_type(),
-                    description(),
-                    #[cfg(feature = "pgp")]
-                    encrypt(),
-                    #[cfg(feature = "pgp")]
-                    sign(),
-                ))
-                .repeated()
-                .collect::<Props>(),
-            )
-            .then_ignore(just(GREATER_THAN))
-            .then_ignore(new_line().or_not())
-            .then(
-                choice((
-                    multipart,
-                    single_part(),
-                    plain_text_part(1).map(Part::PlainText),
-                ))
-                .repeated()
-                .collect(),
-            )
-            .then_ignore(just(MULTI_PART_END))
-            .then_ignore(new_line().or_not())
-            .map(|(props, parts)| Part::Multi(props, parts))
-    })
+/// It parses everything that is inside and outside (multi)parts.
+pub(crate) fn plain_text_part<'a>(
+    min: usize,
+) -> impl Parser<'a, &'a str, &'a str, ParserError<'a>> + Clone {
+    any()
+        .and_is(choice((part_begin(), part_end(), multipart_begin(), multipart_end())).not())
+        .repeated()
+        .at_least(min)
+        .slice()
 }
 
 #[cfg(test)]
@@ -142,7 +109,7 @@ mod parts {
     #[test]
     fn single_part_no_new_line() {
         assert_eq!(
-            super::single_part()
+            super::part()
                 .parse("<#part>This is a plain text part.")
                 .into_result(),
             Ok(Part::Single(
@@ -152,7 +119,7 @@ mod parts {
         );
 
         assert_eq!(
-            super::single_part()
+            super::part()
                 .parse("<#part>This is a plain text part.<#/part>")
                 .into_result(),
             Ok(Part::Single(
@@ -165,7 +132,7 @@ mod parts {
     #[test]
     fn single_part_new_line() {
         assert_eq!(
-            super::single_part()
+            super::part()
                 .parse(concat_line!("<#part>", "This is a plain text part."))
                 .into_result(),
             Ok(Part::Single(
@@ -175,7 +142,7 @@ mod parts {
         );
 
         assert_eq!(
-            super::single_part()
+            super::part()
                 .parse(concat_line!(
                     "<#part>",
                     "This is a plain text part.",
@@ -193,7 +160,7 @@ mod parts {
     #[test]
     fn single_html_part() {
         assert_eq!(
-            super::single_part()
+            super::part()
                 .parse(concat_line!(
                     "<#part type=text/html>",
                     "<h1>This is a HTML text part.</h1>",
@@ -210,7 +177,7 @@ mod parts {
     #[test]
     fn attachment() {
         assert_eq!(
-            super::single_part()
+            super::part()
                 .parse("<#part type=image/jpeg filename=~/rms.jpg disposition=inline><#/part>")
                 .into_result(),
             Ok(Part::Single(
@@ -227,7 +194,7 @@ mod parts {
     #[test]
     fn multi_part() {
         assert_eq!(
-            super::multi_part()
+            super::multipart()
                 .parse(concat_line!(
                     "<#multipart>",
                     "This is a plain text part.",
@@ -244,7 +211,7 @@ mod parts {
     #[test]
     fn nested_multi_part() {
         assert_eq!(
-            super::multi_part()
+            super::multipart()
                 .parse(concat_line!(
                     "<#multipart>",
                     "<#multipart>",
@@ -263,7 +230,7 @@ mod parts {
         );
 
         assert_eq!(
-            super::multi_part()
+            super::multipart()
                 .parse(concat_line!(
                     "<#multipart>",
                     "<#multipart>",
@@ -295,7 +262,7 @@ mod parts {
     #[test]
     fn adjacent_multi_part() {
         assert_eq!(
-            super::multi_part()
+            super::multipart()
                 .parse(concat_line!(
                     "<#multipart>",
                     "<#multipart>",