+++ mml/src/message/body/mod.rs
@@ -18,14 +18,15 @@ pub use self::compiler::MmlBodyCompiler;
 #[doc(inline)]
 pub use self::interpreter::{FilterParts, MimeBodyInterpreter};
 
-pub(crate) const SINGLE_PART_BEGIN: &str = "<#part";
-pub(crate) const SINGLE_PART_BEGIN_ESCAPED: &str = "<#!part";
-pub(crate) const SINGLE_PART_END: &str = "<#/part>";
-pub(crate) const SINGLE_PART_END_ESCAPED: &str = "<#!/part>";
-pub(crate) const MULTI_PART_BEGIN: &str = "<#multipart";
-pub(crate) const MULTI_PART_BEGIN_ESCAPED: &str = "<#!multipart";
-pub(crate) const MULTI_PART_END: &str = "<#/multipart>";
-pub(crate) const MULTI_PART_END_ESCAPED: &str = "<#!/multipart>";
+pub(crate) const PART_BEGIN: &str = "<#part";
+pub(crate) const PART_BEGIN_ESCAPED: &str = "<#!part";
+pub(crate) const PART_END: &str = "<#/part>";
+pub(crate) const PART_END_ESCAPED: &str = "<#!/part>";
+
+pub(crate) const MULTIPART_BEGIN: &str = "<#multipart";
+pub(crate) const MULTIPART_BEGIN_ESCAPED: &str = "<#!multipart";
+pub(crate) const MULTIPART_END: &str = "<#/multipart>";
+pub(crate) const MULTIPART_END_ESCAPED: &str = "<#!/multipart>";
 
 pub(crate) const ALTERNATIVE: &str = "alternative";
 pub(crate) const ATTACHMENT: &str = "attachment";